+++ packages/iocuak-e2e/src/definitions/common/models/worlds/TagWorld.ts
@@ -1,6 +1,6 @@
-import { BindingTag } from '@cuaklabs/iocuak';
+import { Tag } from '@cuaklabs/iocuak-common';
 import { IWorld } from '@cucumber/cucumber';
 
 export interface TagWorld extends IWorld {
-  tag: BindingTag;
+  tag: Tag;
 }