+++ packages/iocuak/src/binding/models/api/BaseBindingApi.ts
@@ -1,10 +1,9 @@
-import { ServiceId } from '@cuaklabs/iocuak-common';
+import { ServiceId, Tag } from '@cuaklabs/iocuak-common';
 
-import { BindingTag } from '../domain/BindingTag';
 import { BindingTypeApi } from './BindingTypeApi';
 
 export interface BaseBindingApi {
   id: ServiceId;
   bindingType: BindingTypeApi;
-  tags: BindingTag[];
+  tags: Tag[];
 }