+++ packages/iocuak/src/binding/models/api/InjectableOptionsApi.ts
@@ -1,10 +1,9 @@
-import { ServiceId } from '@cuaklabs/iocuak-common';
+import { ServiceId, Tag } from '@cuaklabs/iocuak-common';
 
-import { BindingTag } from '../domain/BindingTag';
 import { BindingScopeApi } from './BindingScopeApi';
 
 export interface InjectableOptionsApi {
   id?: ServiceId;
   scope?: BindingScopeApi;
-  tags?: BindingTag | BindingTag[];
+  tags?: Tag | Tag[];
 }