+++ packages/iocuak/src/binding/models/domain/BaseBinding.ts
@@ -1,10 +1,9 @@
-import { ServiceId } from '@cuaklabs/iocuak-common';
+import { ServiceId, Tag } from '@cuaklabs/iocuak-common';
 
-import { BindingTag } from './BindingTag';
 import { BindingType } from './BindingType';
 
 export interface BaseBinding {
   id: ServiceId;
   bindingType: BindingType;
-  tags: BindingTag[];
+  tags: Tag[];
 }