+++ packages/iocuak/src/binding/models/domain/BindingTag.ts
@@ -1,1 +1,0 @@
-export type BindingTag = number | string | symbol;