+++ packages/iocuak/src/container/services/api/ContainerInstanceServiceApi.ts
@@ -1,10 +1,6 @@
-import { ServiceId } from '@cuaklabs/iocuak-common';
-
-import { BindingTag } from '../../../binding/models/domain/BindingTag';
+import { ServiceId, Tag } from '@cuaklabs/iocuak-common';
 
 export interface ContainerInstanceServiceApi {
   get<TInstance>(serviceId: ServiceId): TInstance;
-  getByTag<TInstances extends unknown[] = unknown[]>(
-    tag: BindingTag,
-  ): TInstances;
+  getByTag<TInstances extends unknown[] = unknown[]>(tag: Tag): TInstances;
 }