+++ packages/iocuak/src/container/utils/bindToValue.ts
@@ -1,13 +1,12 @@
-import { ServiceId } from '@cuaklabs/iocuak-common';
+import { ServiceId, Tag } from '@cuaklabs/iocuak-common';
 
-import { BindingTag } from '../../binding/models/domain/BindingTag';
 import { BindingType } from '../../binding/models/domain/BindingType';
 import { ValueBinding } from '../../binding/models/domain/ValueBinding';
 import { BindingService } from '../../binding/services/domain/BindingService';
 
 export function bindToValue<TInstance>(
   serviceId: ServiceId,
-  tags: BindingTag[],
+  tags: Tag[],
   value: TInstance,
   containerBindingService: BindingService,
 ): void {