+++ packages/iocuak/CHANGELOG.md
@@ -19,6 +19,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 
+## [UNRELEASED]
+
+### Changed
+- Updated dependencies require `reflect-metadata` as dev dependency.
+- Update dependencies with `@cuaklabs/iocuak-` internal libraries.
+
+
+
+
 ## 0.3.1 - 2022-07-09
 
 ### Changed