+++ packages/iocuak/CHANGELOG.md
@@ -21,6 +21,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [UNRELEASED]
 
+
+
+
+## 0.4.0 - 2022-12-28
+
 ### Added
 - Added `BindOptions`.